* commit 'a86f8810e83f61b64d6659912e731293e5dce5ba':
  QUAL-251: Only load driver if necessary; check ens3
  QUAL-30: Fix typos in script
  QUAL-30: More tweaks plus logging to work on the timing
  QUAL-30: Adjust timeouts for slow IFE tools
  QUAL-30: Adjust debug output that's spamming the guest side
  QUAL-30: WIP changes to sequencing based on testing
  QUAL-30: Update menu title

+++ src/systemd/qual-ife.sh
@@ -2,14 +2,19 @@
 echo "Qual package version is: `rpm -q --queryformat='%{VERSION}' qual-ife`"
 
 # Install and load Microchip driver for I2C devices
-if ! driver_install.sh; then
-    echo "Failed to install Microchip driver"
-elif ! driver_load.sh; then
-    echo "Failed to load Microchip driver"
+lsmod | fgrep -q i2c_mcp2221
+if [ $? != 0 ]; then
+    if ! driver_install.sh; then
+        echo "Failed to install Microchip driver"
+    elif ! driver_load.sh; then
+        echo "Failed to load Microchip driver"
+    fi
 fi
 
 # Set up device node that CLI tools expect
-ln -s i2c-1 /dev/i2c-7
+if [ ! -e /dev/i2c-7 ]; then
+    ln -s i2c-1 /dev/i2c-7
+fi
 
 # Make GPIO FTDI_CONN_CTRL_I2C_SEL = low
 # to set i2c mux to be controlled by x86 processor
@@ -20,6 +25,13 @@ fi
 # Enable all I2C devices through I2C switch
 if ! i2c 7 0x73 0x00 0x7f > /dev/null 2>&1; then
     echo "Failed to enable i2c devices"
+fi
+
+# Set up host-guest virtual network interface
+ifconfig ens3 | fgrep -q 192.168.122.
+if [ $? != 0 ]; then
+    sleep 1
+    ifup ens3
 fi
 
 # Start Qual Test Application
